requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cowsay==6.0
databases==0.9.0
distlib==0.3.8
Django==5.0
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-restframework==0.0.1
djangorestframework==3.15.1
fastapi==0.115.14
filelock==3.13.3
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
Pillow==10.1.0
platformdirs==4.2.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pypng==0.20220715.0
python-dotenv==1.1.1
python-multipart==0.0.20
qrcode==7.4.2
requests==2.32.4
segno==1.6.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.4.4
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.25.1
yarl==1.20.1
